Add a python file that creates input for java
The functions used to create the data files can easily by read
by the InputReader class in java.
The first two lines of the file refer to the size of the array
and the size of the query array.
The query array can be looped over to find the average time
taken by the implementations.